edit: BottomSheet -> SortModalViewController
